refactor(oracle): migrate igmp to per-protocol plugins
